tea优化

+++ app/Controller/Admin/TeaController.php
@@ -123,7 +123,7 @@ class TeaController extends AdminController
                 $profit=new TeaProfit();
                 $profit->user_count=(new TeaUser())->value('count(*)');
                 $profit->received=math(4980,$profit->user_count,'*',2);
-                $profit->support=(new TeaMoney())->value('sum(money)');
+                $profit->support=(float)(new TeaMoney())->value('sum(money)');
                 $profit->rate=math($profit->support,$profit->received,'/',5);
                 $profit->save();
